refactor: create base component for a generic scrollable dialog, use in DuplicateDialog
